+++ packages/renderer/src/renderer/screens/OpenGov/Referenda/ReferendumRow.tsx
@@ -28,8 +28,8 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 import { TooltipRx } from '@polkadot-live/ui/components';
 import { FlexRow } from '@polkadot-live/ui/styles';
-import { ReferendumDropdownMenu } from '../DropdownMenu';
-import { RoundLeftButton, RoundRightButton } from '../DropdownMenu/Wrappers';
+import { ReferendumDropdownMenu } from '../DropdownMenus';
+import { RoundLeftButton, RoundRightButton } from '../DropdownMenus/Wrappers';
 import type { RefStatus } from '@polkadot-live/types/openGov';
 import type { ReferendumRowProps } from '../types';
